Actualização

+++ Gestão_de_Sistema_Bancário/Clientes.cs
@@ -8,8 +8,9 @@ namespace Gestão_de_Sistema_Bancário
 {
     class Clientes : Pessoa
     {
-        public int Contacto { get; set; }
-        public Clientes(int BI, string nome, string sobrenome, DateTime datanascimento, int bi, string sexo, int contacto, int IdMunicipio) : base (BI, nome, sobrenome, datanascimento, bi, sexo, IdMunicipio)
+        
+        public long Contacto { get; set; }
+        public Clientes(int BI, string nome, string sobrenome, DateTime datanascimento, int bi, string sexo, long contacto, int IdMunicipio) : base (BI, nome, sobrenome, datanascimento, bi, sexo, IdMunicipio)
         {
             this.Contacto = contacto;
         }
